requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.119.1
h11==0.16.0
idna==3.11
pydantic==2.12.3
pydantic-core==2.41.4
sniffio==1.3.1
starlette==0.48.0
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0
sqlalchemy==2.0.32
aiosqlite==0.18.0
python-dotenv==1.0.0
databases
pydantic-settings
rich
asgi-correlation-id
python-json-logger
logtail-python
python-jose
python-multipart
passlib
bcrypt==4.3.0
